refactor(api): split admin-roles mutations into per-file endpoints (ADR 0001)
Split useAddAdminMutation, useRemoveAdminMutation, and
useUpdateAdminRoleMutation out of useAdminRolesQuery.ts into their own
self-contained files per ADR 0001 (one use* file per endpoint). The
query file now holds only the query; consumer imports repointed. No
behavior change.
